backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
pillow==10.1.0
python-dotenv==1.0.0
httpx==0.25.2
bcrypt==4.1.1
alembic==1.12.1
